+++ lib/database_consistency/checkers/validators_fraction_checkers/numericality_constraint_checker.rb
@@ -21,7 +21,8 @@ module DatabaseConsistency
       def preconditions
         model.connection.respond_to?(:check_constraints) &&
           model.connection.table_exists?(model.table_name) &&
-          column
+          column &&
+          validators.any?
       end
 
       def check