+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### [Unreleased]
+
+- Fix `NumericalityConstraintChecker` false positives for attributes that have validators but no `numericality` validator. Thanks [panteo](https://github.com/panteo) for reporting this in [#309](https://github.com/djezzzl/database_consistency/issues/309)!
+
 ### [3.0.7] - 2026/07/26
 
 - Add `NumericalityConstraintChecker` to detect `numericality` validators without matching DB `CHECK` constraints. Thanks [Chedli Bourguiba](https://github.com/chaadow) for reporting this in [#181](https://github.com/djezzzl/database_consistency/issues/181)!